requirements.txt
chromadb
httpx
aiofiles
sqlalchemy
numpy
dataflows
kvfile
awesome-slugify
